+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix LSP completion for options.
 - Add LSP document highlighting support.
 - Add LSP completion for fully-qualified type references.
+- Improve LSP semantic tokens implementation (for syntax highlighting).
 
 ## [v1.63.0] - 2026-01-06